removed conversion tests, since not applicable to tensorflow plus not very meaningful
